+++ app/src/main/java/com/fibelatti/photowidget/home/InvalidWidgetBottomSheet.kt
@@ -13,57 +13,45 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.fibelatti.photowidget.R
 import com.fibelatti.photowidget.configure.PhotoWidgetConfigureActivity
-import com.fibelatti.photowidget.model.PhotoWidgetStatus
 import com.fibelatti.photowidget.ui.WarningSign
 import com.fibelatti.ui.foundation.AppSheetState
 import com.fibelatti.ui.foundation.SelectionDialogBottomSheet
 import com.fibelatti.ui.foundation.data
 
 @Composable
-fun RemovedWidgetBottomSheet(
+fun InvalidWidgetBottomSheet(
     sheetState: AppSheetState,
-    onKeep: (appWidgetId: Int) -> Unit,
     onDelete: (appWidgetId: Int) -> Unit,
     modifier: Modifier = Modifier,
 ) {
     val localContext: Context = LocalContext.current
     val localResources: Resources = LocalResources.current
-    val data: RemovedWidgetBottomSheetData = sheetState.data() ?: return
+    val appWidgetId: Int = sheetState.data() ?: return
 
     SelectionDialogBottomSheet(
         sheetState = sheetState,
-        options = buildList {
-            add(RemovedWidgetOptions.RESTORE)
-            if (PhotoWidgetStatus.KEPT != data.status) {
-                add(RemovedWidgetOptions.KEEP)
-            }
-            add(RemovedWidgetOptions.DELETE)
-        },
+        options = InvalidWidgetOptions.entries,
         optionName = { option -> localResources.getString(option.label) },
         onOptionSelected = { option ->
             when (option) {
-                RemovedWidgetOptions.RESTORE -> {
-                    val intent = PhotoWidgetConfigureActivity.restoreWidgetIntent(
+                InvalidWidgetOptions.EDIT -> {
+                    val intent = PhotoWidgetConfigureActivity.editWidgetIntent(
                         context = localContext,
-                        appWidgetId = data.appWidgetId,
+                        appWidgetId = appWidgetId,
                     )
 
                     localContext.startActivity(intent)
                 }
 
-                RemovedWidgetOptions.KEEP -> {
-                    onKeep(data.appWidgetId)
-                }
-
-                RemovedWidgetOptions.DELETE -> {
-                    onDelete(data.appWidgetId)
+                InvalidWidgetOptions.DELETE -> {
+                    onDelete(appWidgetId)
                 }
             }
         },
         modifier = modifier,
         footer = {
             WarningSign(
-                text = stringResource(id = R.string.photo_widget_home_removed_widgets_hint),
+                text = stringResource(R.string.photo_widget_home_invalid_widgets_hint),
                 modifier = Modifier
                     .fillMaxWidth()
                     .padding(all = 16.dp),
@@ -72,16 +60,10 @@ fun RemovedWidgetBottomSheet(
     )
 }
 
-class RemovedWidgetBottomSheetData(
-    val appWidgetId: Int,
-    val status: PhotoWidgetStatus,
-)
-
-private enum class RemovedWidgetOptions(
+private enum class InvalidWidgetOptions(
     @StringRes val label: Int,
 ) {
 
-    RESTORE(label = R.string.photo_widget_home_removed_widget_action_restore),
-    KEEP(label = R.string.photo_widget_home_removed_widget_action_keep),
+    EDIT(label = R.string.photo_widget_home_my_widget_action_edit),
     DELETE(label = R.string.photo_widget_home_removed_widget_action_delete),
 }